splited some files, but there is a problem
