+++ Components/Pages/Counter.razor
@@ -1,24 +1,8 @@
-@page "/counter"
 @rendermode InteractiveServer
-@using Syncfusion.Blazor.Buttons
+@page "/counter"
 
-<PageTitle>Counter</PageTitle>
 <h1>Counter</h1>
-<SfButton @ref="ToggleBtn" @onclick="onToggleClick" CssClass="e-flat" IsToggle="true" IsPrimary="true"
-    Content="@Content"></SfButton>
-@code {
-    public string Content = "Play";
-    SfButton? ToggleBtn;
-    @* 问题: is it a good way to solve the problem from nullable. *@
-    private void onToggleClick(Microsoft.AspNetCore.Components.Web.MouseEventArgs args)
-    {
-        if (Content == "Play")
-        {
-            this.Content = "Pause";
-        }
-        else
-        {
-            this.Content = "Play";
-        }
-    }
-}
+
+<p>Current count: @currentCount</p>
+
+<button class="btn btn-primary" @onclick="IncrementCount">Click me</button>
