+++ Components/Pages/Counter.razor.cs
@@ -1,0 +1,9 @@
+public partial class Counter
+{
+    private int currentCount = 0;
+
+    private void IncrementCount()
+    {
+        currentCount++;
+    }
+}
