+++ Components/Pages/Test.razor.cs
@@ -1,0 +1,11 @@
+namespace Pages{
+public partial class Test
+{
+    private int currentCount = 0;
+
+    private void IncrementCount()
+    {
+        currentCount++;
+    }
+}
+}
