+++ Components/Shared/TableComponent.razor
@@ -1,4 +1,4 @@
-@using static BlazorApp.Components.Shared.DateBlock
+@using Global
 @using static BlazorApp.Components.Shared.TableRow
 @rendermode InteractiveServer
