+++ Components/Shared/TableRow.razor
@@ -1,4 +1,5 @@
 @using static BlazorApp.Components.Shared.DateBlock
+@using Global
 @if (UserData != null)
 {
     <li class="table-row user-row">
